Updated verison of FarmManager
The program now includes the header row that Tea added to the CSV files, skipping the first line and ensuring that the header is written back.
When creating new invoices, the system will prompt the user to determine if the customer wishes to use their points. If yes, you can decide how many points, then those points will be deducted from the user.
Customer selection is now available as a drop-down menu, allowing you to search for a user when creating an invoice. 
There is also a points button on customer i was playing with, so can edit their points either add or remove.